Rework split assignment; Move to f64 over f32

+++ dronalize/datasets/highd/map/builder.py
@@ -49,8 +49,8 @@ class HighDMapBuilder(BaseMapBuilder):
         # These are used implicitly if `MapBuilder.build` is called.
         _min_distance, _interp_distance = min_distance, interp_distance
         data = pl.read_csv(self._meta_file).select(
-            pl.col("upperLaneMarkings").str.split(";").cast(pl.List(pl.Float32)),
-            pl.col("lowerLaneMarkings").str.split(";").cast(pl.List(pl.Float32)),
+            pl.col("upperLaneMarkings").str.split(";").cast(pl.List(pl.Float64)),
+            pl.col("lowerLaneMarkings").str.split(";").cast(pl.List(pl.Float64)),
         )
 
         n_lane_markings = len(data["upperLaneMarkings"][0])
